+++ src/TaskBuildSymbolTree.h
@@ -49,8 +49,12 @@ public:
     void registerActivityLabels(ast::ISymbolScope *scope);
     virtual void visitConstraintBlock(ast::IConstraintBlock *i) override;
 
+    virtual void visitGenericConstraintDeclBool(ast::IGenericConstraintDeclBool *i) override;
+
+    virtual void visitGenericConstraintDeclValue(ast::IGenericConstraintDeclValue *i) override;
+
     virtual void visitConstraintScope(ast::IConstraintScope *i) override;
-    
+
     virtual void visitConstraintStmt(ast::IConstraintStmt *i) override;
 
     virtual void visitConstraintStmtForall(ast::IConstraintStmtForall *i) override;
@@ -73,6 +77,8 @@ public:
 
     virtual void visitExtendType(ast::IExtendType *i) override;
 
+    virtual void visitExtendEnum(ast::IExtendEnum *i) override;
+
     virtual void visitField(ast::IField *i) override;
 
     virtual void visitFieldCompRef(ast::IFieldCompRef *i) override;
@@ -80,6 +86,24 @@ public:
     virtual void visitFieldRef(ast::IFieldRef *i) override;
 
     virtual void visitFieldClaim(ast::IFieldClaim *i) override;
+
+    virtual void visitActionHandleField(ast::IActionHandleField *i) override;
+
+    virtual void visitCovergroup(ast::ICovergroup *i) override;
+
+    virtual void visitCovergroupType(ast::ICovergroupType *i) override;
+
+    virtual void visitCovergroupInstantiation(ast::ICovergroupInstantiation *i) override;
+
+    virtual void visitExportAction(ast::IExportAction *i) override;
+
+    virtual void visitSymbolDeclaration(ast::ISymbolDeclaration *i) override;
+
+    virtual void visitMonitorActivityDecl(ast::IMonitorActivityDecl *i) override;
+
+    virtual void visitCoverStmtInline(ast::ICoverStmtInline *i) override;
+
+    virtual void visitCoverStmtReference(ast::ICoverStmtReference *i) override;
 
     virtual void visitFunctionDefinition(ast::IFunctionDefinition *i) override;