+++ src/TaskResolveRefs.h
@@ -64,7 +64,6 @@ public:
 
     virtual void visitExprRefPathContext(ast::IExprRefPathContext *i) override;
 
-    virtual void visitExprRefPathId(ast::IExprRefPathId *i) override;
 
     virtual void visitExprRefPathStatic(ast::IExprRefPathStatic *i) override;